# swaggerV2.json
{
  "swagger": "2.0",
  "info": {
    "description": "RESTfull API baseada em node.js Com o objetivo da partilha de viagens e experiencias vividas, classificações e comentáriosentre utilizadores. Disponível em https://tppweb.herokuapp.com",
    "version": "2.0.0",
    "title": "Trips Together",
    "contact": {
      "email": "o.nosso.email@gmail.com"
    }
  },
...
  "produces": [
    "application/json"
  ],
  "host": "tppweb.herokuapp.com",
  "basePath": "/api/v2",
  "paths": {
...
    "/classificacao/viagem/{id}": {
      "get": {
        "tags": [
          "Classificação"
        ],
        "summary": "Obter classificações de uma viagem",
        "description": "Obtem todos as classificações a uma determinada viagem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/classificacao/momento/{id}": {
      "get": {
        "tags": [
          "Classificação"
        ],
        "summary": "Obter classificações de um momento",
        "description": "Obtem todos as classificações a um determinado momento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do momento",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/classificacao": {
      "post": {
        "tags": [
          "Classificação"
        ],
        "summary": "Inserir uma viagem",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "classifica",
            "in": "formData",
            "description": "Classificação aplicada",
            "type": "integer",
            "enum": [
              "Já fui muito feliz aqui",
              "Nunca tinha experimentado, estou maravilhado",
              "Este vai direto para a minha bucket list"
            ]
          },
          {
            "name": "idUtilizador",
            "in": "formData",
            "description": "ID do utilizador que classifica a viagem",
            "type": "string"
          },
          {
            "name": "idViagem",
            "in": "formData",
            "description": "ID da viagem a classificar (para o caso de classificar uma viagem)",
            "type": "string"
          },
          {
            "name": "idMomento",
            "in": "formData",
            "description": "ID do momento a classificar (para o caso de classificar um momento)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem inserida com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "o momemto ja foi classificado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/comentarios": {
      "post": {
        "tags": [
          "Comentários"
        ],
        "summary": "Inserir uma viagem",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idMomento",
            "in": "formData",
            "description": "ID do momento (no caso de comentar um momento)",
            "type": "string"
          },
          {
            "name": "idViagem",
            "in": "formData",
            "description": "ID da viagem (no caso de comentar uma viagem)",
            "type": "string"
          },
          {
            "name": "idUtilizador",
            "in": "formData",
            "description": "ID do utilizador",
            "type": "string"
          },
          {
            "name": "comentario",
            "in": "formData",
            "description": "contúdo docomentário",
            "type": "string"
          },
          {
            "name": "_user",
            "in": "formData",
            "description": "ID interno do utilizador",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem inserida com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/comentarios/{id}": {
      "delete": {
        "tags": [
          "Comentários"
        ],
        "summary": "Remover um comentário por id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do comentário",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Comentário removido com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/comentarios/utilizador/{id}": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Obter os comentários de um utilizador",
        "description": "Obtem todos os comentários de um determinado utilizador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/comentarios/viagem/{id}": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Obter os comentários de uma viagem",
        "description": "Obtem todos os comentários de uma viagem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/comentarios/momento/{id}": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Obter os comentários de um momento",
        "description": "Obtem todos os comentários de um determinado momento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do momento",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/localizacao/lat/{lat}/lng/{lng}": {
      "get": {
        "tags": [
          "Localização"
        ],
        "summary": "Obtem a cidade e país através das coordenadas",
        "description": "Com base nas coordenadas são devolvidos o nome da cidade e país correspondente",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "description": "latitude",
            "required": true,
            "type": "string"
          },
          {
            "name": "lng",
            "in": "path",
            "description": "longitude",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/localizacao/cidade/{cidade}/pais/{pais}": {
      "get": {
        "tags": [
          "Localização"
        ],
        "summary": "Obtem coordenadas",
        "description": "Obtem as coordenadas através do nome de uma cidade e um país",
        "parameters": [
          {
            "name": "cidade",
            "in": "path",
            "description": "Cidade",
            "required": true,
            "type": "string"
          },
          {
            "name": "pais",
            "in": "path",
            "description": "Pais",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/localizacao/meteo/lat/{lat}/lng/{lng}": {
      "get": {
        "tags": [
          "Localização"
        ],
        "summary": "Obtem informação metereológica",
        "description": "Obtem informação metereológica sobre a localização das coordenadas inseridas",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "description": "latitude",
            "required": true,
            "type": "string"
          },
          {
            "name": "lng",
            "in": "path",
            "description": "longitude",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/localizacao/meteo/cidade/{cidade}/pais/{pais}": {
      "get": {
        "tags": [
          "Localização"
        ],
        "summary": "Obtem informação metereológica",
        "description": "Obtem informação metereológica sobre a localização com base na cidade e país inseridos",
        "parameters": [
          {
            "name": "cidade",
            "in": "path",
            "description": "Cidade",
            "required": true,
            "type": "string"
          },
          {
            "name": "pais",
            "in": "path",
            "description": "Pais",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Response": {
      "type": "object",
      "properties": {
        "ErroCorde": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        },
        "Error": {
          "type": "object"
        },
        "Data": {
          "type": "object"
        }
      }
    }
  }
}